another example 'book' added, entity hierarchy re-arranged

+++ src/main/java/com/fererlab/city/dao/CityIDAuditedDAO.java
@@ -8,6 +8,6 @@ import javax.ejb.Local;
 
 
 @Local
-public interface CityIDAuditedDAO extends BaseDAO<CityIDAudited, Long> {
+public interface CityIDAuditedDAO extends BaseDAO<CityIDAudited, Integer> {
 
 }
